# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Marble 3D Service API
  description: |
    🎲 Marble 3D Service - 用中文提示词生成 3D 世界

    基于 World Labs API 的智能 3D 场景生成服务，支持：
    - 中文提示词输入（自动翻译成英文并优化）
    - 图片上传生成 3D 世界
    - 本地 LLM 支持（LM Studio / vLLM / llama.cpp / Ollama / 任意 OpenAI 兼容端点）
    - 多后端 3D 生成（Zero123++ / Stable Zero123 / SD img2img）
    - 文生图（SD1.5 / SDXL）
  version: 1.3.0
  contact:
    name: Marble 3D Service Maintainers
    url: https://github.com/cpufreestyle/marble-3d-service/issues

servers:
  - url: http://localhost:5000
    description: 本地开发服务器
  - url: https://your-domain.com
    description: 生产环境

tags:
  - name: World
    description: 3D 世界生成相关接口
  - name: LLM
    description: 本地 LLM 状态检测
  - name: Models
    description: 本地部署模型管理

paths:
  /api/create:
    post:
      tags: [World]
      summary: 创建 3D 世界
      description: |
        创建 3D 世界（支持文字提示词或图片上传）
        必须提供 `prompt` 或 `image` 至少一个。
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                prompt:
                  type: string
                  description: 中文提示词（将自动翻译成英文并优化）
                image:
                  type: string
                  format: binary
                  description: 图片文件（JPG/PNG/WEBP，最大 10MB）
                api_key:
                  type: string
                  description: 自定义 World Labs API Key
                use_local_llm:
                  type: boolean
                  default: true
                engine:
                  type: string
                  enum: [world_labs, stable_3d, auto]
                  default: stable_3d
                  description: 指定生成引擎
                llm_model:
                  type: string
                  description: 指定本地 LLM 模型（见 GET /api/models）
                three_d_backend:
                  type: string
                  enum: [zero123plus, stable-zero123, sd-img2img, auto]
                  description: 指定 3D 生成后端
          application/json:
            schema:
              type: object
              properties:
                prompt:
                  type: string
                image_url:
                  type: string
                  format: url
                api_key:
                  type: string
                use_local_llm:
                  type: boolean
                  default: true
                engine:
                  type: string
                  enum: [world_labs, stable_3d, auto]
                llm_model:
                  type: string
                three_d_backend:
                  type: string
                  enum: [zero123plus, stable-zero123, sd-img2img, auto]
      responses:
        '200':
          description: 创建成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  engine_used:
                    type: string
                    description: 使用的引擎
                  task_id:
                    type: string
                  status:
                    type: string
                    enum: [processing, completed]
                  original_prompt:
                    type: string
                  enhanced_prompt:
                    type: string
                  llm_used:
                    type: string
                  image_url:
                    type: string
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: 缺少 API Key
        '500':
          description: 服务器内部错误

  /api/task/{task_id}:
    get:
      tags: [World]
      summary: 获取任务状态
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: string
        - name: api_key
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/TaskProcessing'
                  - $ref: '#/components/schemas/TaskCompleted'

  /api/llm-status:
    get:
      tags: [LLM]
      summary: 获取本地 LLM 状态
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  available:
                    type: boolean
                  type:
                    type: string
                    description: LLM 提供方（lmstudio/ollama/vllm 等）
                  url:
                    type: string
                    format: url
                  model:
                    type: string
                    description: 当前使用的模型
                  models:
                    type: array
                    items:
                      type: string
                    description: 可用模型列表

  /api/models:
    get:
      tags: [Models]
      summary: 列出可用的本地部署模型
      description: 返回 LLM / 文生图 / 3D 生成后端的可用模型与硬件适配状态
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  llm:
                    type: object
                    properties:
                      available:
                        type: boolean
                      provider:
                        type: string
                      model:
                        type: string
                      models:
                        type: array
                        items:
                          type: string
                  text_to_image:
                    type: object
                    properties:
                      available:
                        type: boolean
                      model:
                        type: string
                      models:
                        type: array
                        items:
                          type: object
                      pipeline:
                        type: string
                      default_params:
                        type: object
                  three_d:
                    type: object
                    properties:
                      available:
                        type: boolean
                      backend:
                        type: string
                      available_backends:
                        type: array
                        items:
                          type: object
                      device:
                        type: string

  /api/generate-image:
    post:
      tags: [World]
      summary: 文生图
      description: 使用本地 Stable Diffusion / SDXL 从文本提示词生成图片
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                prompt:
                  type: string
                  description: 提示词
                model:
                  type: string
                  description: 模型 ID（运行时切换）
                width:
                  type: integer
                  description: 分辨率（256-1024，8 的倍数，默认 512）
                height:
                  type: integer
                num_inference_steps:
                  type: integer
                  description: 推理步数（1-60，默认 20）
      responses:
        '200':
          description: 生成成功
        '400':
          description: 参数校验失败
        '500':
          description: 模型加载或生成失败

  /api/upload-image:
    post:
      tags: [World]
      summary: 上传图片
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                image:
                  type: string
                  format: binary
                  description: 图片文件（JPG/PNG/WEBP，最大 10MB）
      responses:
        '200':
          description: 上传成功
        '400':
          description: 请求参数错误
        '500':
          description: 服务器内部错误

  /api/engine-status:
    get:
      tags: [World]
      summary: 获取3D生成引擎状态
      responses:
        '200':
          description: 查询成功

components:
  schemas:
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
      required: [success, error]

    TaskProcessing:
      type: object
      properties:
        success:
          type: boolean
        status:
          type: string
          enum: [processing]
        progress:
          type: string
      required: [success, status]

    TaskCompleted:
      type: object
      properties:
        success:
          type: boolean
        status:
          type: string
          enum: [completed]
        result:
          type: object
          properties:
            world_id:
              type: string
            world_url:
              type: string
              format: url
            preview_url:
              type: string
              format: url
            pano_url:
              type: string
              format: url
            thumbnail_url:
              type: string
              format: url
            caption:
              type: string
            spz_100k:
              type: string
              format: url
            spz_500k:
              type: string
              format: url
            spz_full:
              type: string
              format: url
            mesh_url:
              type: string
              format: url
      required: [success, status, result]

externalDocs:
  description: GitHub 仓库
  url: https://github.com/cpufreestyle/marble-3d-service